+++ src/main/java/nl/nn/testtool/storage/memory/Storage.java
@@ -91,10 +91,10 @@ public class Storage implements CrudStorage, LogStorage {
 		return new ArrayList(storageIds);
 	}
 
-	public synchronized List getMetadata(int numberOfRecords, List metadataNames,
+	public synchronized List getMetadata(int maxNumberOfRecords, List metadataNames,
 			List searchValues, int metadataValueType) {
 		List result = new ArrayList();
-		for (int i = 0; i < metadata.size() && (numberOfRecords == -1 || i < numberOfRecords); i++) {
+		for (int i = 0; i < metadata.size() && (maxNumberOfRecords == -1 || i < maxNumberOfRecords); i++) {
 			Map metadataRecord = (Map)metadata.get(i);
 			List resultRecord = new ArrayList();
 			Iterator metadataNamesIterator = metadataNames.iterator();
@@ -117,16 +117,6 @@ public class Storage implements CrudStorage, LogStorage {
 			}
 		}
 		return result;
-	}
-
-	public List getTreeChildren(String path) {
-		// TODO implementeren?
-		return new ArrayList();
-	}
-
-	public List getStorageIds(String path) throws StorageException {
-		// TODO implementeren?
-		return new ArrayList();
 	}
 
 	public synchronized Report getReport(Integer storageId) {